Verify cert not expire for xds client as well
