+++ pkg/istio-agent/xds_proxy.go
@@ -714,6 +714,16 @@ func (p *XdsProxy) getTLSDialOption(agent *Agent) (grpc.DialOption, error) {
 			var certificate tls.Certificate
 			key, cert := agent.GetKeyCertsForXDS()
 			if key != "" && cert != "" {
+				var isExpired bool
+				isExpired, err = util.IsCertExpired(cert)
+				if err != nil {
+					log.Warnf("cannot parse the cert chain, using token instead: %v", err)
+					return &certificate, nil
+				}
+				if isExpired {
+					log.Warnf("cert expired, using token instead")
+					return &certificate, nil
+				}
 				// Load the certificate from disk
 				certificate, err = tls.LoadX509KeyPair(cert, key)
 				if err != nil {
