--- security/pkg/nodeagent/caclient/providers/citadel/testdata/token
@@ -1,1 +1,0 @@
-fake
