+++ security/pkg/pki/testdata/notexpired-cert.pem
